requirements.txt
astroid==2.0.4
boardgamegeek2==0.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.7
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
pylint==2.1.1
redis==3.2.1
requests==2.20.0
requests-cache==0.4.13
rq==1.0
six==1.11.0
urllib3==1.24.2
Werkzeug==0.14.1
wrapt==1.10.11
